input sys dir, file list, norm dir
Expanded the command line arguments to no longer be positional
and to inclue --input-systems for the classifier outputs, --input-norm
for the normalization reference file, --file-list for the ordered
file basenames, --input-text for the input plain text, --output-dir
for the output xmi files.

The ordered file list determines which files get processed rather
than the directory contents.

The team names are added with their identification key in the annotation
attribute (discoveryTechnique=0-10 where zero is oracle and 10 is the
last team name) as a note metadata field.