added problematic discontinous example; voting ensemble algorithm
Added an example of a pair of discontinuous annotations that have
the same first begin and last end, which causes a hash clash. Nothing
to do for now but need to fix later.

Voting ensemble algorithm allow you to set minVotes and what to do
when the min threshold isn't met. For this task, CUI-less or drop
are two strategies. I left hooks for implementing other tie-breakers
(including using ranked choice or normalized confidence scores). For
now, we use random CUI of one of the tied systems.