added support for filtering by classifier id
Since the classifiers are all id in discoveryTechnique as 1-10,
we can use that to allow filtering by classifier id.  It's a very
simple system assuming single digit for all ids (10 -> 0 since we know
actual 0 is the oracle). We'll need to beef up the flexibility for
when there are more than 10 classifiers and to make them easily
sync with a classifiers file/ranked file.