New command line arg voting-unit determines focus of vote accumulation
Two of three options for --voting-unit implemented:  'span' and 'sentence'
Document level 'doc' unit not implemented yet.

If the sentence is the voting unit, then we need an external sentence
splitter. Then we iterate over those sentences to seed a span dictionary
(called kb = {} in the current version). Then we iterate over all
annotations and assign counts and weights based on how much overlap
a particular annotation has with a given sentence. That sentence is then
annotated with the concept if some concept exceeds the min vote threshold.
if an annotation doesn't fully cover the sentence, it's weight update is
only 0.5 rather than 1.  Counts are always incremented by 1. No partial
credit.

When the span is the voting unit, logic currently only supports an
exact match of annotation spans.  So exact matching span will have
their votes counted.  Partially overlapping spans are considered to
exist in a different universe.