added classifier rank file, overlap strategy; spans support imperfect matching
The default (and only) overlapping strategy is to use the classifier ranks
to resolve ties.  The ranks are provided in a simple plain text file
via the --classifier-rank-file argument. Higher classifiers are
given preference over lower classifier in the file.

For span annotations, this surfaces as the top classifier being
treated as the base annotation much like the sentence spans
are treated as the base for section annotations. The difference lies
in that if a span from the non-top classifier doesn't overlap with
any of the already seen spans, then a new entry is added to the
voting dictionary for all future classifiers to overlap with.