added context attributes; document level voting units
The two attributes of polarity and uncertainty were added
to the identified annotation type. Plus the term_exists NoteNLP
attribute was re-typed as Boolean (as per the wiki specs).

When the voting unit is 'doc', some of the logic takes a separate
path. As such, I refactored the tallySpannedVotes from the previous
logic of sentence and span-level voting. That leaves tallyDocVotes
as a special function just for document level context attributes
surrounding lists of concepts. Passing the minVote threshold is required
to be added as a concept. Then the majority vote for each attribute
of the listed concepts is used. Ties are awarded to the most interesting
or novel value.  -1 (negated) is more novel than 1 (explicitly present)
is mroe novel than 0 (implicitly present).