fixed bug when no rankFile; initial prep scripts for cuis and spans
When no ranking file is provided, the random "top" classifier is
pulled from the list of classifiers provided as a argument. This
is the ID, so we need to use that directly (rather than having to
look it up in the classifier2id index).

Following the same template used for context attributes, this is
the initial script for CUIs (2019 n2c2 track 3) and spans (2009
medications challenge).