allow the rankFile to be a tsv
Split each line of a tab. If the file is actually a tsv (with later
columns containing performance metrics for later usage), those will
get ignored. If the file is a simple ranked listing, the split on
tabs is a NOOP.

If the first line starts with "Classifier", then this is a header
row and we can skip it.