removed accuracy from scoring; match at doc-property; simplify ref annotation
We don't really need Accuracy report for context-attributes so it
has been removed from the ETUDE calls.

ETUDE uses the new experimental doc-property match flag to allow
for accurate TN metrics to be counted.

The ref annotation was simplified for context-attributes so that
the term_modifier value always contains exactly one key/value
pair for uncertainty, allowing us to score that as well.